09-practical-projects/06-reinforcement-learning/02-dino-run-dqn/requirements.txt
tensorflow>=2.8.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
selenium>=4.0.0
